ceil/floor/truncate/round return integers.
This CL supercedes https://codereview.chromium.org/11748016/

Review URL: https://codereview.chromium.org//12317107

git-svn-id: https://dart.googlecode.com/svn/branches/bleeding_edge/dart@19810 260f80e4-7a28-3924-810f-c04153c831b5

+++ pkg/fixnum/test/int_64_vm_test.dart
@@ -56,7 +56,7 @@ int _randomInt(int n) {
   for (int i = 0; i < DISCARD; i++) {
     rand = random.nextDouble();
   }
-  return (rand * n).floor().toInt();
+  return (rand * n).floor();
 }
 
 class Op {
