+++ pkg/intl/lib/src/date_format_field.dart
@@ -345,7 +345,7 @@ class _DateFormatPatternField extends _DateFormatField {
   }
 
   String formatQuarter(DateTime date) {
-    var quarter = (date.month / 3).truncate().toInt();
+    var quarter = (date.month / 3).truncate();
     if (width < 4) {
       return symbols.SHORTQUARTERS[quarter];
     } else {
